pyproject.toml
[tool.poetry]
name = "phoebesnotebook"
version = "1.0.0"
description = ""
authors = ["Gilead Raab"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
gunicorn = "^20.1.0"
pyyaml = "^6.0"
markupsafe = "^2.1.2"
markdown = "^3.4.1"
pygments = "^2.14.0"
pymdown-extensions = "^9.9.1"


[tool.poetry.group.dev.dependencies]
ansible = "^7.1.0"
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
